+++ tests/Endpoint/EnvelopeRecipientsEndpointTest.php
@@ -71,6 +71,12 @@ class EnvelopeRecipientsEndpointTest extends EndpointTestCase
         self::assertLastRequest('GET', '/api/envelopes/bar/recipients/foo/identifications');
     }
 
+    public function testScenario(): void
+    {
+        self::endpoint()->scenario('foo');
+        self::assertLastRequest('GET', '/api/envelopes/bar/recipients/foo/scenario');
+    }
+
     protected static function endpoint(): EnvelopeRecipientsEndpoint
     {
         return self::dgs()->envelopes()->recipients('bar');